% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{academy}
\alias{academy}
\title{Pixar awards and nominations}
\format{
A data frame with 72 rows and 3 variables:
\describe{
\item{film}{name of film}
\item{award_type}{name of award}
\item{status}{status of award}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Pixar_films}
}
\usage{
academy
}
\description{
A dataset with the awards and nominations each Pixar film has won.
}
\details{
This dataset is put into a tidy format, where each row is a film-award data
point.

Some films did not qualify for categories so there is no data for that film.
}
\keyword{datasets}
