% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genres}
\alias{genres}
\title{Genres describing Pixar films}
\format{
A data frame with 204 rows and 3 variables:
\describe{
\item{film}{name of film}
\item{category}{classification of genre as either "Genre" or "Subgenre"}
\item{value}{value for genre or subgenre film is categorized as}
}
}
\source{
\url{https://www.omdbapi.com/}
}
\usage{
genres
}
\description{
A data set with the different genre categories for each Pixar film.
}
\details{
This data set is put into a tidy format, where each row is a film-genre data
point. Each film can have multiple genres. For example, Toy Story is
categorized into three genres. Moreover, Toy Story has five subgenres.
}
\examples{
genres
unique(genres$value)
}
\keyword{datasets}
