% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pixar_franchises}
\alias{pixar_franchises}
\title{Pixar film franchises}
\format{
A data frame with 17 rows and 2 columns:
\describe{
\item{film}{name of film}
\item{franchise}{name of franchise the film is part of}
}
}
\usage{
pixar_franchises
}
\description{
A data set with the Pixar films that are part of a franchise and the name of
the franchise.
}
\details{
Note, not all Pixar films are part of a franchise. For example, "A Bug's
Life" is not part of a franchise, but "Toy Story 2" is part of the "Toy Story"
franchise.
}
\examples{
pixar_franchises
}
\keyword{datasets}
