% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pixar_rankings}
\alias{pixar_rankings}
\title{Pixar film rankings}
\format{
A data frame with 521 rows and 3 columns:
\describe{
\item{film}{name of film}
\item{source}{name of internet ranking source}
\item{ranking}{numerical ranking of the film}
}
}
\usage{
pixar_rankings
}
\description{
A data set with the rankings of Pixar films from various sources.
}
\details{
The sources include these websites:

\enumerate{
\item Rotten Tomatoes
\item IGN
\item IndieWire
\item Slant
\item Vox
\item WIRED
\item Thrillist
\item ScreenRant
\item Polygon
\item Buzzfeed
\item CNET
\item BusinessInsider
\item AV Club
\item Vulture
\item Independent
\item MensHealth
\item Time
\item Esquire
\item Mashable
}
}
\examples{
pixar_rankings
}
\keyword{datasets}
