% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{themes_vox}
\alias{themes_vox}
\title{Themes in Pixar films by Vox}
\format{
A data frame with 124 rows and 2 columns:
\describe{
\item{film}{name of film}
\item{theme}{theme in film}
}
}
\source{
\url{https://www.vox.com/2015/11/23/9780818/pixar-chart-movies-toy-story-anniversary}
}
\usage{
themes_vox
}
\description{
A data set with the themes in Pixar films as described by Vox.
}
\examples{
themes_vox
}
\keyword{datasets}
