% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{box_office}
\alias{box_office}
\title{Box office reception numbers}
\format{
A data frame with 23 rows and 5 variables:
\describe{
\item{film}{name of film}
\item{budget}{movie budget in U.S. dollars}
\item{box_office_us_canada}{box office gross amount in U.S. dollars for
U.S. and Canada}
\item{box_office_other}{box office gross amount in U.S. dollars for other
territories}
\item{box_office_worldwide}{box office gross amount in U.S. dollars
worldwide}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Pixar_films}
}
\usage{
box_office
}
\description{
A dataset with financial and box office gross numbers for each film.
}
\examples{
box_office
}
\keyword{datasets}
