% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genres}
\alias{genres}
\title{Genres describing Pixar films}
\format{
A data frame with 128 rows and 2 variables:
\describe{
\item{film}{name of film}
\item{genre}{genre film is categorized into}
}
}
\source{
\url{https://www.omdbapi.com/}
}
\usage{
genres
}
\description{
A data set with the different genre categories for each Pixar film.
}
\details{
This data set is put into a tidy format, where each row is a film-genre data
point. Each film can have multiple genres. For example, Toy Story is
categorized into five genres.
}
\examples{
genres
}
\keyword{datasets}
