% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pixar_films}
\alias{pixar_films}
\title{Pixar films}
\format{
A data frame with 31 rows and 6 columns:
\describe{
\item{number}{order of release}
\item{film}{name of film (primary key)}
\item{release_date}{date film premiered}
\item{run_time}{film length in minutes}
\item{film_rating}{rating based on Motion Picture Association (MPA) film
rating system}
\item{plot}{brief description of the movie plot}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Pixar_films}

\url{https://www.omdbapi.com}
}
\usage{
pixar_films
}
\description{
A data set containing Pixar films, their release order, and release date.
}
\details{
Order of release, film name, and release date come from Wikipedia. The rest
of the columns (run time, film rating, and plot) come from OMDb.

The column \code{film} is the primary key that joins with all the rest of the
datasets in this package.
}
\examples{
pixar_films
mean(pixar_films$run_time)
table(pixar_films$film_rating)
}
\keyword{datasets}
