% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pixar_films}
\alias{pixar_films}
\title{Pixar films}
\format{
A data frame with 27 rows and 3 variables:
\describe{
\item{number}{order of release}
\item{film}{name of film}
\item{release_date}{date film premiered}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Pixar_films}
}
\usage{
pixar_films
}
\description{
A data set containing Pixar films, their release order, and release date.
}
\examples{
pixar_films
}
\keyword{datasets}
