% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pixar_people}
\alias{pixar_people}
\title{People behind Pixar}
\format{
A data frame with 286 rows and 3 columns:
\describe{
\item{film}{name of film}
\item{role_type}{one of six roles: Director, Musician, Producer,
Screenwriter, Storywriter, Co-director}
\item{name}{individual's name}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Pixar_films}
}
\usage{
pixar_people
}
\description{
A data set containing the main people involved in each Pixar film.
}
\details{
This data set is put into a tidy format, where each row is a film-role data
point. Each film can have multiple individuals per role. For example, Toy
Story has four screenwriters.
}
\examples{
pixar_people
}
\keyword{datasets}
