% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pixar_people}
\alias{pixar_people}
\title{People behind Pixar}
\format{
A data frame with 228 rows and 3 variables:
\describe{
\item{film}{name of film}
\item{role_type}{one of five roles: Director, Musician, Producer,
Screenwriter, Storywriter}
\item{name}{individual's name}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Pixar_films}
}
\usage{
pixar_people
}
\description{
A dataset containing the main people involved in each Pixar film.
}
\details{
This dataset is put into a tidy format, where each row is a film-role data
point. Each film can have multiple individuals per role. For example, Toy
Story has four screenwriters.
}
\keyword{datasets}
