% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{public_response}
\alias{public_response}
\title{Critical and public response}
\format{
A data frame with 23 rows and 5 variables:
\describe{
\item{film}{name of film}
\item{rotten_tomatoes_score}{score from the American review-aggregation
website Rotten Tomatoes; scored out of 100}
\item{rotten_tomatoes_counts}{number of critics contributing to Rotten
Tomatoes score}
\item{metacritic_score}{score from Metacritic where scores are weighted
average of reviews; scored out of 100}
\item{metacritic_counts}{number of critics contributing to Metacritic
score}
\item{cinema_score}{score from market research firm CinemaScore; scored by
grades A, B, C, D, and F}
\item{imdb_score}{score from IMDb where scores are weighted average of
reviews; scored out of 100}
\item{imdb_counts}{number of critics contributing to IMDb score}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Pixar_films}
}
\usage{
public_response
}
\description{
A data set with scores of critical and public response.
}
\examples{
public_response
}
\keyword{datasets}
