% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{public_response}
\alias{public_response}
\title{Critical and public response}
\format{
A data frame with 23 rows and 5 variables:
\describe{
\item{film}{name of film}
\item{rotten_tomatoes}{score from the American review-aggregation website
Rotten Tomatoes; scored out of 100}
\item{metacritic}{score from Metacritic where scores are weighted average
of reviews; scored out of 100}
\item{cinema_score}{score from market research firm CinemaScore; scored by
grades A, B, C, D, and F}
\item{critics_choice}{score from Critics' Choice Movie Awards presented by
the American-Canadian Critics Choice Association (CCA); scored out of
100}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Pixar_films}
}
\usage{
public_response
}
\description{
A dataset with scores of critical and public response.
}
\examples{
public_response
}
\keyword{datasets}
